perf(langgraph): avoid iterating on channels if no managed values are present
Cuts down the sequential(3000) from 12s to 8s.
