+++ .changeset/good-plums-film.md
@@ -1,0 +1,5 @@
+---
+"@langchain/langgraph": patch
+---
+
+Avoid iterating on channels if no managed values are present
