+++ libs/langgraph/src/channels/base.ts
@@ -108,21 +108,37 @@ export abstract class BaseChannel<
   }
 }
 
+const IS_ONLY_BASE_CHANNEL = Symbol.for("LG_IS_ONLY_BASE_CHANNEL");
+export function getOnlyChannels(
+  channels: Record<string, BaseChannel>
+): Record<string, BaseChannel> {
+  // @ts-expect-error - we know it's a record of base channels
+  if (channels[IS_ONLY_BASE_CHANNEL] === true) return channels;
+
+  const newChannels = {} as Record<string, BaseChannel>;
+  for (const k in channels) {
+    if (!Object.prototype.hasOwnProperty.call(channels, k)) continue;
+    const value = channels[k];
+    if (isBaseChannel(value)) newChannels[k] = value;
+  }
+
+  Object.assign(newChannels, { [IS_ONLY_BASE_CHANNEL]: true });
+  return newChannels;
+}
+
 export function emptyChannels<Cc extends Record<string, BaseChannel>>(
   channels: Cc,
   checkpoint: ReadonlyCheckpoint
 ): Cc {
-  const filteredChannels = Object.fromEntries(
-    Object.entries(channels).filter(([, value]) => isBaseChannel(value))
-  ) as Cc;
+  const filteredChannels = getOnlyChannels(channels) as Cc;
 
   const newChannels = {} as Cc;
   for (const k in filteredChannels) {
-    if (Object.prototype.hasOwnProperty.call(filteredChannels, k)) {
-      const channelValue = checkpoint.channel_values[k];
-      newChannels[k] = filteredChannels[k].fromCheckpoint(channelValue);
-    }
+    if (!Object.prototype.hasOwnProperty.call(filteredChannels, k)) continue;
+    const channelValue = checkpoint.channel_values[k];
+    newChannels[k] = filteredChannels[k].fromCheckpoint(channelValue);
   }
+  Object.assign(newChannels, { [IS_ONLY_BASE_CHANNEL]: true });
   return newChannels;
 }
 
@@ -138,7 +154,7 @@ export function createCheckpoint<ValueType>(
     values = checkpoint.channel_values;
   } else {
     values = {};
-    for (const k of Object.keys(channels)) {
+    for (const k in channels) {
       try {
         values[k] = channels[k].checkpoint();
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
