+++ libs/langgraph/src/pregel/index.ts
@@ -32,7 +32,7 @@ import {
   BaseChannel,
   createCheckpoint,
   emptyChannels,
-  isBaseChannel,
+  getOnlyChannels,
 } from "../channels/base.js";
 import {
   CHECKPOINT_NAMESPACE_END,
@@ -2042,13 +2042,7 @@ export class Pregel<
       config?.runName ?? this.getName() // run_name
     );
 
-    const channelSpecs: Record<string, BaseChannel> = {};
-    for (const [name, spec] of Object.entries(this.channels)) {
-      if (isBaseChannel(spec)) {
-        channelSpecs[name] = spec;
-      }
-    }
-
+    const channelSpecs = getOnlyChannels(this.channels);
     let loop: PregelLoop | undefined;
     let loopError: unknown;
