+++ libs/langgraph/src/tests/utils.ts
@@ -1,17 +1,8 @@
 /* eslint-disable no-promise-executor-return */
 /* eslint-disable import/no-extraneous-dependencies */
 import { expect, it } from "vitest";
-import { v4 as uuidv4 } from "uuid";
-import { CallbackManagerForLLMRun } from "@langchain/core/callbacks/manager";
 import { Graph as DrawableGraph } from "@langchain/core/runnables/graph";
 import {
-  BaseChatModel,
-  BaseChatModelParams,
-  BaseChatModelCallOptions,
-  BindToolsInput,
-} from "@langchain/core/language_models/chat_models";
-import {
-  BaseMessage,
   AIMessage,
   HumanMessage,
   BaseMessageFields,
@@ -22,8 +13,7 @@ import {
   FunctionMessage,
   FunctionMessageFieldsWithName,
 } from "@langchain/core/messages";
-import { ChatGenerationChunk, ChatResult } from "@langchain/core/outputs";
-import { RunnableConfig, RunnableLambda } from "@langchain/core/runnables";
+import { RunnableConfig } from "@langchain/core/runnables";
 import { Tool } from "@langchain/core/tools";
 import {
   MemorySaver,
@@ -35,285 +25,10 @@ import {
 } from "@langchain/langgraph-checkpoint";
 import { z } from "zod";
 import { BaseTracer, Run } from "@langchain/core/tracers/base";
-import {
-  BaseLanguageModelCallOptions,
-  BaseLanguageModelInput,
-  LanguageModelLike,
-} from "@langchain/core/language_models/base";
 import { Pregel, PregelInputType, PregelOutputType } from "../pregel/index.js";
 import { StrRecord } from "../pregel/algo.js";
 import { PregelNode } from "../pregel/read.js";
 import { BaseChannel, LangGraphRunnableConfig } from "../web.js";
-
-export interface FakeChatModelArgs extends BaseChatModelParams {
-  responses: BaseMessage[];
-}
-
-export class FakeChatModel extends BaseChatModel {
-  responses: BaseMessage[];
-
-  callCount = 0;
-
-  streamMessageId: "omit" | "first-only" | "always";
-
-  constructor(
-    fields: FakeChatModelArgs & {
-      streamMessageId?: "omit" | "first-only" | "always";
-    }
-  ) {
-    super(fields);
-    this.responses = fields.responses;
-    this.streamMessageId = fields.streamMessageId ?? "omit";
-  }
-
-  _combineLLMOutput() {
-    return [];
-  }
-
-  _llmType(): string {
-    return "fake";
-  }
-
-  async _generate(
-    messages: BaseMessage[],
-    options?: this["ParsedCallOptions"]
-  ): Promise<ChatResult> {
-    if (options?.stop?.length) {
-      return {
-        generations: [
-          {
-            message: new AIMessage(options.stop[0]),
-            text: options.stop[0],
-          },
-        ],
-      };
-    }
-    const response = this.responses[this.callCount % this.responses.length];
-    const text = messages.map((m) => m.content).join("\n");
-    this.callCount += 1;
-    return {
-      generations: [
-        {
-          message: response ?? new AIMessage(text),
-          text: response ? (response.content as string) : text,
-        },
-      ],
-      llmOutput: {},
-    };
-  }
-
-  async *_streamResponseChunks(
-    _input: BaseLanguageModelInput,
-    _options?: BaseLanguageModelCallOptions,
-    runManager?: CallbackManagerForLLMRun
-  ) {
-    const response = this.responses[this.callCount % this.responses.length];
-
-    let isFirstChunk = true;
-    const completionId = response.id ?? uuidv4();
-
-    for (const content of (response.content as string).split("")) {
-      let id: string | undefined;
-      if (
-        this.streamMessageId === "always" ||
-        (this.streamMessageId === "first-only" && isFirstChunk)
-      ) {
-        id = completionId;
-      }
-
-      const chunk = new ChatGenerationChunk({
-        message: new AIMessageChunk({ content, id }),
-        text: content,
-      });
-
-      await runManager?.handleLLMNewToken(
-        content,
-        undefined,
-        undefined,
-        undefined,
-        undefined,
-        { chunk }
-      );
-
-      // TODO: workaround for the issue found in Node 18.x
-      // where @langchain/core/utils/stream AsyncGeneratorWithSetup
-      // does for some reason not yield the first chunk to the consumer
-      // and instead the LLM token callback is seen first.
-      yield chunk;
-
-      isFirstChunk = false;
-    }
-    this.callCount += 1;
-  }
-}
-
-export class FakeToolCallingChatModel extends BaseChatModel {
-  sleep?: number = 50;
-
-  responses?: BaseMessage[];
-
-  thrownErrorString?: string;
-
-  idx: number;
-
-  toolStyle: "openai" | "anthropic" | "bedrock" | "google" = "openai";
-
-  structuredResponse?: Record<string, unknown>;
-
-  // Track messages passed to structured output calls
-  structuredOutputMessages: BaseMessage[][] = [];
-
-  constructor(
-    fields: {
-      sleep?: number;
-      responses?: BaseMessage[];
-      thrownErrorString?: string;
-      toolStyle?: "openai" | "anthropic" | "bedrock" | "google";
-      structuredResponse?: Record<string, unknown>;
-    } & BaseChatModelParams
-  ) {
-    super(fields);
-    this.sleep = fields.sleep ?? this.sleep;
-    this.responses = fields.responses;
-    this.thrownErrorString = fields.thrownErrorString;
-    this.idx = 0;
-    this.toolStyle = fields.toolStyle ?? this.toolStyle;
-    this.structuredResponse = fields.structuredResponse;
-    this.structuredOutputMessages = [];
-  }
-
-  _llmType() {
-    return "fake";
-  }
-
-  async _generate(
-    messages: BaseMessage[],
-    _options: this["ParsedCallOptions"],
-    runManager?: CallbackManagerForLLMRun
-  ): Promise<ChatResult> {
-    if (this.thrownErrorString) {
-      throw new Error(this.thrownErrorString);
-    }
-    if (this.sleep !== undefined) {
-      await new Promise((resolve) => setTimeout(resolve, this.sleep));
-    }
-    const responses = this.responses?.length ? this.responses : messages;
-    const msg = responses[this.idx % responses.length];
-    const generation: ChatResult = {
-      generations: [
-        {
-          text: "",
-          message: msg,
-        },
-      ],
-    };
-    this.idx += 1;
-
-    if (typeof msg.content === "string") {
-      await runManager?.handleLLMNewToken(msg.content);
-    }
-    return generation;
-  }
-
-  bindTools(tools: BindToolsInput[]) {
-    const toolDicts = [];
-    const serverTools = [];
-    for (const tool of tools) {
-      if (!("name" in tool)) {
-        serverTools.push(tool);
-        continue;
-      }
-
-      // NOTE: this is a simplified tool spec for testing purposes only
-      if (this.toolStyle === "openai") {
-        toolDicts.push({
-          type: "function",
-          function: {
-            name: tool.name,
-          },
-        });
-      } else if (["anthropic", "google"].includes(this.toolStyle)) {
-        toolDicts.push({
-          name: tool.name,
-        });
-      } else if (this.toolStyle === "bedrock") {
-        toolDicts.push({
-          toolSpec: {
-            name: tool.name,
-          },
-        });
-      }
-    }
-    let toolsToBind: BindToolsInput[] = toolDicts;
-    if (this.toolStyle === "google") {
-      toolsToBind = [{ functionDeclarations: toolDicts }];
-    }
-    return this.bind({
-      tools: [...toolsToBind, ...serverTools],
-    } as BaseChatModelCallOptions);
-  }
-
-  withStructuredOutput<
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    RunOutput extends Record<string, any> = Record<string, any>
-  >(_: unknown) {
-    if (!this.structuredResponse) {
-      throw new Error("No structured response provided");
-    }
-    // Create a runnable that returns the proper structured format
-    return RunnableLambda.from(async (messages: BaseMessage[]) => {
-      if (this.sleep) {
-        await new Promise((resolve) => setTimeout(resolve, this.sleep));
-      }
-
-      // Store the messages that were sent to generate structured output
-      this.structuredOutputMessages.push([...messages]);
-
-      // Return in the format expected: { raw: BaseMessage, parsed: RunOutput }
-      return this.structuredResponse as RunOutput;
-    });
-  }
-}
-
-export class FakeConfigurableModel extends BaseChatModel {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  _queuedMethodOperations: Record<string, any> = {};
-
-  _chatModel: LanguageModelLike;
-
-  constructor(
-    fields: {
-      model: LanguageModelLike;
-    } & BaseChatModelParams
-  ) {
-    super(fields);
-    this._chatModel = fields.model;
-  }
-
-  _llmType() {
-    return "fake_configurable";
-  }
-
-  async _generate(
-    _messages: BaseMessage[],
-    _options: this["ParsedCallOptions"],
-    _runManager?: CallbackManagerForLLMRun
-  ): Promise<ChatResult> {
-    throw new Error("Not implemented");
-  }
-
-  async _model() {
-    return this._chatModel;
-  }
-
-  bindTools(tools: BindToolsInput[]) {
-    const modelWithTools = new FakeConfigurableModel({
-      model: (this._chatModel as FakeToolCallingChatModel).bindTools(tools),
-    });
-    modelWithTools._queuedMethodOperations.bindTools = tools;
-    return modelWithTools;
-  }
-}
 
 export class MemorySaverAssertImmutable extends MemorySaver {
   storageForCopies: Record<string, Record<string, Uint8Array>> = {};
